consolidating client vm setup

+++ client/init.sh
@@ -1,7 +1,17 @@
 #! /bin/bash -e
 
-wget -O $HOME/miniconda.sh https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh
-bash $HOME/miniconda.sh -b -p $HOME/conda
-$HOME/conda/bin/conda update -n base -y conda
-$HOME/conda/bin/conda init
-rm $HOME/miniconda.sh
+cmd="$1"
+shift
+
+if [[ "$cmd" == "install" ]]; then
+    wget -O $HOME/miniconda.sh https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh
+    bash $HOME/miniconda.sh -b -p $HOME/conda
+    $HOME/conda/bin/conda update -n base -y conda
+    $HOME/conda/bin/conda init
+    rm $HOME/miniconda.sh
+elif [[ "$cmd" == "create" ]]; then
+    conda env create -f environment.yaml
+elif [[ "$cmd" == "run" ]]; then
+    conda activate o2
+    python client.py "$@"
+fi
